backend/Scrawler/requirements1.txt
apturl==0.5.2
beautifulsoup4==4.5.3
blinker==1.3
Brlapi==0.6.4
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
checkbox-support==0.22
click==6.7
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
defer==1.0.6
dill==0.2.7.1
feedparser==5.1.3
firebase-admin==2.9.0
Flask==0.12.2
Flask-CLI==0.4.0
future==0.16.0
google-api-core==0.1.4
google-auth==1.4.1
google-cloud-core==0.28.0
google-cloud-firestore==0.28.0
google-cloud-storage==1.7.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.9.1
guacamole==0.9.2
gunicorn==19.7.1
html5lib==0.999
httplib2==0.9.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==1.0
matlab==0.1
matplotlib==2.0.2
numpy==1.13.3
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pbr==3.1.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
ply==3.8
protobuf==3.5.1
ptyprocess==0.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-engineio==1.7.0
python-firebase==1.2
python-socketio==1.8.1
python-systemd==231
pytz==2018.3
pyxdg==0.25
reportlab==3.3.0
requests==2.13.0
rsa==3.4.2
sessioninstaller==0.0.0
six==1.11.0
stevedore==1.27.1
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.22
usb-creator==0.3.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
Werkzeug==0.14.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
